Add in API keys and a simple permission model (#17)
* Add in API keys and a simple permission model

I've added in the required table and model for API keys. Currently they
have a simple permission model, each key can have any combination of the
following boolean permissions:

> **read:** view all repo data
> **write:** manually run new builds, purge, maybe others (not implemented)
> **admin:** grant and revoke access through API keys (not implemented)

There's also a middleware named `requireAuth` which can be used
per-route to enforce the permissions.

* Switch to require an API key and secret pair

I've updated to make endpoints require an API key and secret pair. This
means that even if the database were compromised, there would be no need
for key rotation and we could safely restore to a previous backup.

The secrets are hashed using the `bcrypt` module. We reduce the number
of salt rounds required in order to speed up the service. Because of
this compromise, we should never use UUIDs as secrets: these could be
guessable.

* Add a method to prepare for generating secret keys