Normalise semver versions in the database (#24)
This allows for easier querying later. We need to keep both the
non-normalised and normalised versions for different reasons:

We'll use the non-normalised version to fetch details about a GitHub
release. For this, we need to match the tag on GitHub _exactly_.

We'll use the normalised version to power semver-based URLs and so that
database rows can be sorted by this field reliably. E.g. we don't end up
with the following:

  - 1.0.0
  - 1.3.0
  - v1.2.0