Process ingestions and gather data from GitHub (#38)
Now the ingestion queue processor gathers data from GitHub and creates
Version entities from Ingestion entities. There's a lot of code here,
but it wasn't really possible to split much more than this and have
something functional in place.

If everything goes well, an Ingestion will be converted into a Version
when all of the available information is fetched from GitHub. If an
error occurs:

  1. We determine whether the error is recoverable or not. An example of
    a recoverable error is if the rate limit has been exceeded. An
    example of a non-recoverable error is if the repo/tag does not
    include and Origami manifest.

    1. For recoverable errors we increment the number of attempts and will
       try again later.

    2. For non-recoverable errors we remove the ingestion and do not
       attempt again.

I've also added logic to prevent an Ingestion from being created if it
already exists or if it's already been converted to a Version. This
reduces our GitHub API usage.

Running locally this has ingested 500-600 versions per hour, it would be
faster if we didn't keep hitting the GitHub API rate limit. I'm not
going to address this as we're highly unlikely to have this many
releases on GitHub in any given hour.

I haven't added in cleanup of over-attempted or long-running ingestions
yet, this will come in a separate PR. I've also not found a nice way to
run integration tests on this part of the code-base, in the interest of
brevity for this already-large set of changes, I'll add tests in a
separate PR as well.