Use 307 redirects and redocument them as "aliases" (#44)
We need to use 307 redirects as they allow forwarding on of the request
method and headers. This is important for us because authentication
details are sent as headers.

I've also redocumented the redirecting endpoint as an "alias" of the one
it redirects to. I think this will aid discoverability and reduce the
length of the documentation somewhat, as we were repeating a lot.