Sort semantic versions correctly (#57)
We have not been sorting semantic versions correctly, this work fixes
that. We now split the versions before storing each part in the database
so we can query against them and sort them appropriately.

There's a costly migration in here which I've tested against a copy of
the live database so we know that it runs fine.

I've exposed the raw git tag for a version, because when build metadata
is present this is stripped from the sanitized semantic version. This
means that a consuming client could not link directly to the tag.

I've also exposed our calculation of whether a version is a prerelease
in the output for repos and versions so that this logic doesn't have to
be duplicated in a consuming client.